Adapt to new App Lifecycle changes in Hoist
