+++ client-app/src/apps/admin.js
@@ -1,4 +1,16 @@
 import {XH} from '@xh/hoist/core';
-import {App} from '../admin/App';
 
-XH.renderApp(App);
+import {App} from '@xh/hoist/admin/App';
+import {AppModel} from '@xh/hoist/admin/AppModel';
+import {AppContainer} from '@xh/hoist/desktop/appcontainer';
+
+XH.renderApp({
+    componentClass: App,
+    modelClass: AppModel,
+    containerClass: AppContainer,
+    isMobile: false,
+    isSSO: false,
+    checkAccess: ['HOIST_ADMIN'],
+    loginMessage: 'Contact support@xh.io for information on Hoist\'s bundled Admin Console.',
+    idleDetectionDisabled: true
+});
