+++ client-app/src/apps/app.js
@@ -1,4 +1,14 @@
 import {XH} from '@xh/hoist/core';
 import {App} from '../desktop/App';
+import {AppModel} from '../desktop/AppModel';
+import {AppContainer} from '@xh/hoist/desktop/appcontainer';
 
-XH.renderApp(App);
+XH.renderApp({
+    componentClass: App,
+    modelClass: AppModel,
+    containerClass: AppContainer,
+    isMobile: false,
+    isSSO: false,
+    checkAccess: ['APP_READER'],
+    loginMessage: "User: 'toolbox@xh.io' / Password: 'toolbox'"
+});
