+++ client-app/src/apps/mobile.js
@@ -1,4 +1,15 @@
 import {XH} from '@xh/hoist/core';
 import {App} from '../mobile/App';
+import {AppModel} from '../mobile/AppModel';
+import {AppContainer} from '@xh/hoist/mobile/appcontainer';
 
-XH.renderApp(App);
+
+XH.renderApp({
+    componentClass: App,
+    modelClass: AppModel,
+    containerClass: AppContainer,
+    isMobile: true,
+    isSSO: false,
+    checkAccess: ['APP_READER'],
+    loginMessage: "User: 'toolbox@xh.io' / Password: 'toolbox'"
+});
